Support for non-working copy repositories
Sometimes we just want a repository on a hard drive somewhere but we don't want
to duplicate every file by also having a working copy sitting around. In this
case, to help prevent someone from thinking that there is nothing in a specific
directory, we don't want to use the "." prefix for the layout and objects
directories.